+++ src/djhtmx/__init__.py
@@ -1,4 +1,4 @@
 from .middleware import middleware
 
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 __all__ = ("middleware",)